refactored libs
