+++ src/lib/is-in-dir.js
@@ -1,0 +1,11 @@
+const isInDir = (path, dirs) => {
+  for (let i = 0, l = dirs.length; i < l; i++) {
+    if (dirs[i].indexOf(path) === 0) {
+      return true
+    }
+  }
+
+  return false
+}
+
+export default isInDir
